🎨 refactor: 정렬 로직 수정
